@@ -1,22 +1,14 @@
 import React from 'react';
-import ReactDOM from 'react-dom';
+import {createRoot} from 'react-dom/client';
+import './index.css';
 import App from './App';
-import * as serviceWorker from './serviceWorker';
-import { Router } from 'react-router-dom';
-import { createBrowserHistory } from 'history';
+import reportWebVitals from './reportWebVitals';
 
-const history = createBrowserHistory();
+const container = document.getElementById("app");
+const root = createRoot(container);
+root.render(<App tab="home" />);
 
-ReactDOM.render(
-  <React.StrictMode>
-    <Router history={history}>
-      <App history={history} />
-    </Router>
-  </React.StrictMode>,
-  document.getElementById('root')
-);
-
-// If you want your app to work offline and load faster, you can change
-// unregister() to register() below. Note this comes with some pitfalls.
-// Learn more about service workers: https://bit.ly/CRA-PWA
-serviceWorker.unregister();
+// If you want to start measuring performance in your app, pass a function
+// to log results (for example: reportWebVitals(console.log))
+// or send to an analytics endpoint. Learn more: https://bit.ly/CRA-vitals
+reportWebVitals();
