@@ -1,2 +1,3 @@
-export const toFirstCharUppercase = name =>
-    name.charAt(0).toUpperCase() + name.slice(1);
+export const toFirstCharUppercase = (name) => {
+    return name.charAt(0).toUpperCase() + name.slice(1);
+}
