@@ -1,22 +1,113 @@
-const typeColors = {
-    bug: '#729f3f',
-    dragon: '#53a4cf',
-    fairy: '#fdb9e9',
-    fire: '#fd7d24',
-    ghost: '#7b62a3',
-    ground: '#f7de3f',
-    normal: '#a4acaf',
-    pyschic: '#f366b9',
-    steel: '#9eb7b',
-    dark: '#707070',
-    electric: '#eed535',
-    fighting: '#d56723',
-    flying: '#3dc7ef',
-    grass: '#9bcc50',
-    ice: '#51c4e7',
-    poison: '#b97fc9',
-    rock: '#a38c21',
-    water: '#4592c4'
-}
+export const typeColors = {
+  normal: "#9199A1",
+  fighting: "#CE4069",
+  flying: "#92AADE",
+  poison: "#AB6AC8",
+  ground: "#D97746",
+  rock: "#C7B78B",
+  bug: "#90C12C",
+  ghost: "#5269AC",
+  steel: "#5A8EA1",
+  fire: "#FF9C54",
+  water: "#4E90D5",
+  electric: "#F3D23B",
+  grass: "#63BB5B",
+  ice: "#74CEC0",
+  psychic: "#F97176",
+  dragon: "#0A6DC4",
+  dark: "#5A5366",
+  fairy: "#EC8FE6",
+};
 
-export default typeColors;
+export const typeList = [
+{
+  "name": "normal",
+  "color": "#9199A1",
+  "icon": require("../images/types/Pokemon_Type_Icon_Normal.png")
+},
+{
+  "name": "fighting",
+  "color": "#CE4069",
+  "icon": require("../images/types/Pokemon_Type_Icon_Fighting.png")
+},
+{
+  "name": "flying",
+  "color": "#92AADE",
+  "icon": require("../images/types/Pokemon_Type_Icon_Flying.png")
+},
+{
+  "name": "poison",
+  "color": "#AB6AC8",
+  "icon": require("../images/types/Pokemon_Type_Icon_Poison.png")
+},
+{
+  "name": "ground",
+  "color": "#D97746",
+  "icon": require("../images/types/Pokemon_Type_Icon_Ground.png")
+},
+{
+  "name": "rock",
+  "color": "#C7B78B",
+  "icon": require("../images/types/Pokemon_Type_Icon_Rock.png")
+},
+{
+  "name": "bug",
+  "color": "#90C12C",
+  "icon": require("../images/types/Pokemon_Type_Icon_Bug.png")
+},
+{
+  "name": "ghost",
+  "color": "#5269AC",
+  "icon": require("../images/types/Pokemon_Type_Icon_Ghost.png")
+},
+{
+  "name": "steel",
+  "color": "#5A8EA1",
+  "icon": require("../images/types/Pokemon_Type_Icon_Steel.png")
+},
+{
+  "name": "fire",
+  "color": "#FF9C54",
+  "icon": require("../images/types/Pokemon_Type_Icon_Fire.png")
+},
+{
+  "name": "water",
+  "color": "#4E90D5",
+  "icon": require("../images/types/Pokemon_Type_Icon_Water.png")
+},
+{
+  "name": "grass",
+  "color": "#63BB5B",
+  "icon": require("../images/types/Pokemon_Type_Icon_Grass.png")
+},
+{
+  "name": "electric",
+  "color": "#F3D23B",
+  "icon": require("../images/types/Pokemon_Type_Icon_Electric.png")
+},
+{
+  "name": "psychic",
+  "color": "#F97176",
+  "icon": require("../images/types/Pokemon_Type_Icon_Psychic.png")
+},
+{
+  "name": "ice",
+  "color": "#74CEC0",
+  "icon": require("../images/types/Pokemon_Type_Icon_Ice.png")
+},
+{
+  "name": "dragon",
+  "color": "#0A6DC4",
+  "icon": require("../images/types/Pokemon_Type_Icon_Dragon.png")
+},
+{
+  "name": "dark",
+  "color": "#5A5366",
+  "icon": require("../images/types/Pokemon_Type_Icon_Dark.png")
+},
+{
+  "name": "fairy",
+  "color": "#EC8FE6",
+  "icon": require("../images/types/Pokemon_Type_Icon_Fairy.png")
+},
+]
